+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 - Add Connect service
 - Add MediaTailor support
+- Add ByteStream struct to core
 
 ## [0.35.0] - 2018-10-31